Add Update and Delete
Add update and delete feature to CURD test
